+++ assets/js/player-character/objects/custom-elements/ability-score/AbilityScoreInput.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom input element for editing an ability score.
