+++ assets/js/player-character/objects/custom-elements/custom-objects/CustomObjectDeleteDialog.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom HTML element for displaying the CustomObject Delete Dialog.
