+++ assets/js/player-character/objects/custom-elements/custom-objects/NewCustomObjectButton.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 import { ApiObjectInfo } from "../../api/resources/ApiObjectInfo.js";
 
 export class NewCustomObjectButton extends HTMLButtonElement {
