+++ assets/js/player-character/objects/custom-elements/equipment-proficiencies/ArmorProficiencyCheckbox.js
@@ -1,5 +1,5 @@
 import { Armor } from "../../api/resources/equipment/Armor.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom checkbox element representing an armor proficiency toggle.
