+++ assets/js/player-character/objects/custom-elements/generic-info/BackgroundInput.js
@@ -1,6 +1,6 @@
 import { Background } from "../../api/resources/Background.js";
 import { getEmptyOption, getSelectOption } from "../../../util.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom select element for choosing a character background.
