+++ assets/js/player-character/objects/custom-elements/generic-info/ClassLevelSection.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 import { ClassLevelInput } from "./ClassLevelInput.js";
 
 /**
