+++ assets/js/player-character/objects/custom-elements/generic-info/RaceInput.js
@@ -1,6 +1,6 @@
 import { Race } from "../../api/resources/Race.js";
 import { getEmptyOption, getSelectOption } from "../../../util.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom select element for choosing a race.
