+++ assets/js/player-character/objects/custom-elements/inventory/InventoryArmorEffectiveArmorClassCell.js
@@ -1,5 +1,5 @@
 import { Armor } from "../../api/resources/equipment/Armor.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom table cell element that displays an armor's effective armor class.
