+++ assets/js/player-character/objects/custom-elements/inventory/InventoryWeaponAbilityCell.js
@@ -1,6 +1,6 @@
 import { Weapon } from "../../api/resources/equipment/Weapon.js";
 import { getSelectOption } from "../../../util.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom table cell element that displays or allows selection of the ability used for weapon attacks.
