+++ assets/js/player-character/objects/custom-elements/inventory/InventoryWeaponDamageCell.js
@@ -1,5 +1,5 @@
 import { Weapon } from "../../api/resources/equipment/Weapon.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom table cell element that displays the damage dice and damage bonus of a weapon.
