+++ assets/js/player-character/objects/custom-elements/player-character-bank/CharacterDeleteDialog.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom HTML element for displaying the Character Delete Dialog.
