+++ assets/js/player-character/objects/custom-elements/player-character-bank/CharacterSelectButton.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom HTML element for handling the button to switch between characters.
