+++ assets/js/player-character/objects/custom-elements/player-character-bank/NewCharacterButton.js
@@ -1,4 +1,4 @@
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 import { PlayerCharacter } from "../../PlayerCharacter.js";
 
 /**
