+++ assets/js/player-character/objects/custom-elements/subrace/SubraceFeaturesDisplay.js
@@ -1,6 +1,6 @@
 import { Race } from "../../api/resources/Race.js";
 import { Subrace } from "../../api/resources/Subrace.js";
-import { globals } from "../../../load-page.js";
+import { globals } from "../../../load-globals.js";
 
 /**
  * Custom details element that displays the features of the selected subrace.
