+++ pc-builder/custom-object/index.md
@@ -7,6 +7,8 @@ On this page, you can build your own custom objects. A custom object in this ins
 
 <h2 is="custom-object-form-heading"></h2>
 
+<script type="module" src="{{ '/assets/js/player-character/load-globals.js' | relative_url }}"></script>
+
 <script type="module">
     import { loadPage } from "{{ '/assets/js/player-character/load-page.js' | relative_url }}";
     loadPage();
